move TestTrackState into helper header for reuse
